sprint18/docker-lamertoxa/library/requirements.txt
Django == 3.2.16
PyJWT == 2.4.0
asgiref == 3.4.1
djangorestframework == 3.14.0
djangorestframework-simplejwt == 4.4.0
psycopg2 == 2.9.3
pytz == 2022.2.1
setuptools == 28.8.0
sqlparse == 0.4.2
typing-extensions == 4.1.1
tzdata == 2022.2
django-crispy-forms
